refactor(tools): extract registerReviewBundleTools into tools/review-bundles.js
Phase C item 4: moves preview_review_bundle and create_review_bundle out of
index.js. resolveOutputDirWithinSync is passed through toolContext (closes
over SYNC_DIR_REAL/SYNC_DIR_ABS). Removes REVIEW_BUNDLE_PROFILES,
REVIEW_BUNDLE_STRICTNESS, buildReviewBundlePlan, createReviewBundleArtifacts,
and getHeadCommitHash from index.js imports; ReviewBundlePlanError is retained
as resolveOutputDirWithinSync still uses it.
